♻️ separate image panel into two classes: default and rounded
