+++ murico-swingapp/src/main/java/com/github/hanzm_10/murico/swingapp/ui/components/panels/ImagePanel.java
@@ -13,75 +13,35 @@
  */
 package com.github.hanzm_10.murico.swingapp.ui.components.panels;
 
-import java.awt.AlphaComposite;
-import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.Graphics;
-import java.awt.Graphics2D;
 import java.awt.Image;
-import java.awt.RenderingHints;
-import java.awt.geom.RoundRectangle2D;
-import java.awt.image.BufferedImage;
 
 import javax.swing.JPanel;
 
-import org.jetbrains.annotations.NotNull;
-
 public class ImagePanel extends JPanel {
-	private Image image;
-    private Image roundedImage;
-    private int arc;
+	protected Image image;
 
 	public ImagePanel(Image image) {
 		super();
+
 		this.image = image;
 
+		var width = image.getWidth(null);
+		var height = image.getHeight(null);
+
 		setOpaque(false);
-        var width = image.getWidth(null);
-        var height = image.getHeight(null);
 		setPreferredSize(new Dimension(width, height));
 		setSize(new Dimension(width, height));
 	}
-
-    public int getArc() {
-        return arc;
-    }
-
-    public void setArc(@NotNull final int arc) {
-        this.arc = arc;
-        makeImageRounded();
-    }
-
-    protected void makeImageRounded() {
-        var w = image.getWidth(null);
-        var h = image.getHeight(null);
-
-        var output = new BufferedImage(w, h, BufferedImage.TYPE_INT_ARGB);
-        Graphics2D g2 = output.createGraphics();
-
-        g2.setComposite(AlphaComposite.Src);
-        g2.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
-        g2.setColor(Color.WHITE);
-        g2.fill(new RoundRectangle2D.Float(0, 0, w, h, arc, arc));
-
-        g2.setComposite(AlphaComposite.SrcAtop);
-        g2.drawImage(image, 0, 0, null);
-
-        g2.dispose();
-
-        roundedImage = output;
-    }
 
 	@Override
 	protected void paintComponent(Graphics g) {
 		super.paintComponent(g);
 
 		if (image != null) {
-            var image = roundedImage == null ? this.image : roundedImage;
-
 			var panelWidth = getWidth();
 			var panelHeight = getHeight();
-
 			var imgWidth = image.getWidth(null);
 			var imgHeight = image.getHeight(null);
 
@@ -103,8 +63,10 @@ public class ImagePanel extends JPanel {
 
 	public void setImage(Image image) {
 		this.image = image;
+
 		setPreferredSize(new Dimension(image.getWidth(this), image.getHeight(this)));
-		revalidate();
+		setSize(new Dimension(image.getWidth(this), image.getHeight(this)));
+
 		repaint();
 	}
 }
